Use optional setting instead

+++ simple-dyno.js
@@ -37,11 +37,9 @@ var doc = undefined;
 exports.doc = doc;
 
 function setConfig() {
-  var options = arguments.length <= 0 || arguments[0] === undefined ? { accessKeyId: "", secretAccessKey: "", region: "eu-west-1" } : arguments[0];
+  var options = arguments.length <= 0 || arguments[0] === undefined ? { accessKeyId: "", secretAccessKey: "", region: "eu-west-1", runLocal: true } : arguments[0];
 
-  if (process.env.NODE_ENV === 'travis') {
-    options.endpoint = new AWS.Endpoint('http://localhost:8000');
-  } else if (_debug2['default'].local || process.env.NODE_ENV === 'test') {
+  if ((_debug2['default'].local || process.env.NODE_ENV === 'test') && options.runLocal) {
     try {
       var stdout = (0, _child_process.execSync)("killall java");
       logger('killall java: ' + stdout.split('\n').join(''));
@@ -50,6 +48,8 @@ function setConfig() {
     localDynamo.launch(null, 4567);
     options.endpoint = new AWS.Endpoint('http://localhost:4567');
     logger("Started local DynamoDB on http://localhost:4567");
+  } else if (_debug2['default'].local || process.env.NODE_ENV === 'test') {
+    options.endpoint = new AWS.Endpoint('http://localhost:8000');
   }
 
   exports.client = client = new AWS.DynamoDB(options);
